+++ test/runtests.jl
@@ -17,7 +17,7 @@ include("testmst.jl")
 include("testpmedian.jl")
 include("testcpm.jl")
 include("testpert.jl")
-include("testsimplex.jl")
 include("testlatex.jl")
 include("testknapsack.jl")
+include("testsimplex.jl")